Added Job Result schema as file

+++ topchef/models.py
@@ -4,6 +4,7 @@ have JSON representations written in marshmallow, and a single representation
 in the database.
 """
 import os
+import shutil
 import tempfile
 import uuid
 import logging
@@ -27,25 +28,132 @@ BASE = declarative_base(metadata=database.METADATA)
 
 
 class SchemaDirectoryOrganizer(object):
-    REGISTRATION_SCHEMA_NAME = 'job_registration_schema'
-    RESULT_SCHEMA_NAME = 'job_result_schema'
+    """
+    Directory manager responsible for organizing the directory
+    where services and jobs can store their required JSON Schemas.
+
+    The top directory that is to be managed is passed into the constructor.
+    Under normal execution, this directory is given by ``SCHEMA_DIRECTORY``
+    in :mod:`config.py`.
+
+    Each service is given a directory based on the service ID.
+    Each job is given a directory based on the job ID under the service
+    ID. 
+    
+    For a service ``1`` and a job ``2``. The directory tree will look like
+
+    ::
+
+        /schema_directory
+         |
+         |__/1
+         |   |
+         |   |__/2
+
+    The SchemaDirectoryOrganizer also acts as a repository for constants
+    for the model classes that have to do with writing the models to their
+    required schema directories
+
+    :var str root_path: The name of the top directory that this manager
+        is to organize.
+    """
+    REGISTRATION_SCHEMA_NAME = 'job_registration_schema.json'
+    RESULT_SCHEMA_NAME = 'job_result_schema.json'
+    
+    JOB_PARAMETER_FILE_NAME = 'parameters.json'
+    JOB_RESULT_FILE_NAME = 'result.json'
 
     def __init__(self, schema_directory_path):
+        """
+        Instantiates the variables listed in the class description
+        """
         self.root_path = schema_directory_path
 
     @property
     def services(self):
+        """
+        Returns the id of all services that are registered with this manager.
+        The IDs are retrieved by crawling through the directory.
+
+        :return: A list of UUIDs corresponding to the IDs of all services
+            registered with an instance of this directory manager.
+        :rtype: list(UUID)
+        """
         return [
-            service_id for service_id in os.listdir(self.root_path)
+            UUID(service_id) for service_id in os.listdir(self.root_path)
                 if self._is_guid(service_id)
             ]
 
-    def register_service(self, service):
-        service_directory = os.path.join(self.root_path, str(service.id))
+    def register(self, model):
+        """
+        Register a model class with this manager. Creates a directory
+        for the required model class.
 
-        os.mkdir(service_directory)
+        :param model: The model class to register with this manager
+        :type model: :class:`topchef.models.Job` | 
+            :class:`topchef.models.Service`
+        :raises: ValueError if an invalid model is registered
+        """
+        if isinstance(model, Service):
+            self._register_service(model)
+        elif isinstance(model, Job):
+            self._register_job(model)
+        else:
+            raise ValueError("Attempted to register an invalid model class")
+
+    def _register_service(self, service):
+        """
+        Register a service
+        
+        :raises: ValueError if the service directory already exists
+        """
+        service_path = os.path.join(self.root_path, str(service.id))
+        
+        if os.path.isdir(service_path):
+            raise ValueError("Attempted to register service %s. \
+                    A directory for this service already exists at %s" % (
+                    service, service_path)
+                  )
+        else:
+            os.mkdir(service_path)
+
+        print(service_path)
+        assert os.path.isdir(service_path)
+
+    def _register_job(self, job):
+        """
+        Register a job
+
+        :raises: ValueError if the job directory exists or the job's service
+            directory doesn't exist
+        """
+        service_path = os.path.join(self.root_path, str(job.parent_service.id))
+
+        if not os.path.isdir(service_path):
+            raise ValueError("Attempted to register job %s. \
+                    The job's service %s has no service directory. \
+                    Expected service directory to be %s." % (
+                    job, job.parent_service, service_path)
+            )
+
+        job_path = os.path.join(service_path, str(job.id))
+        
+        if os.path.isdir(job_path):
+            raise ValueError("Attempted to register job %s. \
+                    A directory for this job already exists at %s." % (
+                    job, job_path)
+            )
+        else:
+            os.mkdir(job_path)
 
     def __getitem__(self, model):
+        """
+        Return the directory where each model is located
+
+        :param model: The model class for which the directory must be found
+        :return: The path to the working directory for the model
+        :rtype: str
+        """
         if isinstance(model, Service):
             return os.path.join(
                 self.root_path, str(model.id)
@@ -53,7 +161,7 @@ class SchemaDirectoryOrganizer(object):
         elif isinstance(model, Job):
             return os.path.join(
                 self.root_path, str(model.parent_service.id),
-                '%s.json' % str(model.id)
+                str(model.id)
             )
         else:
             raise ValueError(
@@ -62,17 +170,26 @@ class SchemaDirectoryOrganizer(object):
             )
 
     def write(self, data_to_write, target_path):
-
-        _, temporary_filename = tempfile.mkstemp(suffix='.json')
+        """
+        Write the required data to the target path
+        """
+        file_descriptor, temporary_filename = tempfile.mkstemp(suffix='.json')
 
         with open(temporary_filename, mode='w') as temporary_file:
             temporary_file.write(data_to_write)
 
+        os.close(file_descriptor)
+        
         if os.path.isfile(target_path):
             os.remove(target_path)
+        
+        if not os.path.isdir(os.path.split(target_path)[0]):
+            raise FileNotFoundError("The parent directory %s to which this "
+            "file is to be written does not exist" % os.path.split(target_path)[0])
 
-        os.rename(temporary_file.name, target_path)
+        shutil.move(temporary_filename, target_path)
 
+        
         if os.path.isfile(temporary_filename):
             os.remove(temporary_filename)
 
@@ -127,12 +244,12 @@ class Service(BASE):
         self.name = name
         self.description = description
         self.heartbeat_timeout = heartbeat_timeout
-        self.file_manager = organizer
 
+        self.file_manager = organizer
+        self.file_manager.register(self)
+        
         self.last_checked_in = datetime.utcnow()
         self.is_available = True
-
-        self.file_manager.register_service(self)
 
         if job_registration_schema is None:
             self.job_registration_schema = {'type': 'object'}
@@ -213,7 +330,7 @@ class Service(BASE):
     def job_registration_schema(self):
         registration_schema_path = os.path.join(
             self.file_manager[self],
-            '%s.json' % self.file_manager.REGISTRATION_SCHEMA_NAME
+            self.file_manager.REGISTRATION_SCHEMA_NAME
         )
 
         with open(registration_schema_path, mode='r') as schema_file:
@@ -225,7 +342,7 @@ class Service(BASE):
     def job_registration_schema(self, schema_to_write):
         schema_path = os.path.join(
             self.file_manager[self],
-            '%s.json' % self.file_manager.REGISTRATION_SCHEMA_NAME
+            self.file_manager.REGISTRATION_SCHEMA_NAME
         )
 
         JSONSchema().validate(schema_to_write)
@@ -250,8 +367,7 @@ class Service(BASE):
         if errors: raise ValueError('The supplied schema is not a JSON Schema')
 
         schema_path = os.path.join(
-            self.file_manager[self],
-            '%s.json' % self.file_manager.RESULT_SCHEMA_NAME
+            self.file_manager[self], self.file_manager.RESULT_SCHEMA_NAME
         )
 
         self.file_manager.write(data, schema_path)
@@ -319,14 +435,17 @@ class Job(BASE):
             job_parameters,
             self.parent_service.job_registration_schema
         )
-
+ 
         self.id = uuid.uuid1()
         self.date_submitted = datetime.utcnow()
         self.status = "REGISTERED"
+        
         self.file_manager = file_manager
+        self.file_manager.register(self)
+
         self.session = attached_session
         self.parameters = job_parameters
-
+        
     def __next__(self):
         job = self.session.query(self.__class__).filter(
             self.__class__.date_submitted > self.date_submitted
@@ -352,10 +471,12 @@ class Job(BASE):
 
     @property
     def parameters(self):
-        schema_path = self.file_manager[self]
+        schema_path = os.path.join(
+            self.file_manager[self], self.file_manager.JOB_PARAMETER_FILE_NAME
+        )
 
         if not os.path.isfile(schema_path):
-            with open(schema_path, mode='w') as file:
+            with open(schema_path, mode='w+') as file:
                 file.write(json.dumps({}))
 
         with open(schema_path, mode='r') as file:
@@ -365,11 +486,13 @@ class Job(BASE):
 
     @parameters.setter
     def parameters(self, new_schema):
-        schema_path = self.file_manager[self]
+        schema_path = os.path.join(
+            self.file_manager[self], self.file_manager.JOB_PARAMETER_FILE_NAME
+        )
 
         JSONSchema().validate(new_schema)
 
-        with open(schema_path, mode='w') as schema_file:
+        with open(schema_path, mode='w+') as schema_file:
             schema_file.write(json.dumps(new_schema))
 
     @property
@@ -378,16 +501,26 @@ class Job(BASE):
 
     @property
     def result(self):
-        with open(self.file_manager[self]) as result_file:
+        schema_path = os.path.join(
+            self.file_manager[self], 
+            self.file_manager.JOB_RESULT_FILE_NAME
+        )
+
+        with open(schema_path) as result_file:
             file_data = result_file.read()
 
         return JSONSchema().loads(file_data).data
 
     @result.setter
     def result(self, job_result):
+        path_to_write = os.path.join(
+            self.file_manager[self],
+            self.file_manager.JOB_RESULT_FILE_NAME
+        )
+
         jsonschema.validate(job_result, self.parent_service.job_result_schema)
         self.file_manager.write(
-            json.dumps(job_result), self.file_manager[self]
+            json.dumps(job_result), path_to_write
         )
 
     def __iter__(self):
